refactor: move experiment into new folder
